requirements.txt
fastapi>=0.111.0,<0.115.0
sqlalchemy>=2.0.25,<2.1.0
pydantic>=2.6.0,<2.10.0
pydantic-settings>=2.2.0,<2.5.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<1.8.0
python-multipart>=0.0.6,<0.1.0
alembic>=1.13.1,<1.14.0
uvicorn[standard]>=0.29.0,<0.32.0
python-dotenv>=1.0.0,<2.0.0
aiofiles>=23.2.1,<24.0.0
email-validator==2.1.0